requirements.txt
requests==2.25.1
mypy==0.812
black==20.8b1
isort==5.7.0
python-dateutil==2.8.1
